Fix Gemini's ModuleAblator->ModuleHeatshield patch
@MODULE[...]{} don't work in a new PART{}, they just get
spit out into the MM cache as-is. Move the patch into its
own @PART; make sure it runs before DeadlyReentry's own
convert-and-change-values patch runs.

This restores Gemini to the AdvanceCommandPod-level shield
performance it was intended to have. There's a good chance that's
TOO good (only 31 units consumed on lunar reentry, in a recent
test); if so, the values should be revisited for both capsules
(and perhaps MH Gemini)